separated versions from constants
