+++ src/utils/constants.ts
@@ -94,206 +94,6 @@ export const BADGE_COLORS = {
 
 export const MAX_LENGTH = 30;
 
-export const VERSION_MAP = {
-  "1.3.0": {
-    version: "1.3.0",
-    title: "Select Actions",
-    desc: "Easily select and manage your group members.",
-    updates: [
-      "Increased loading performances in group page",
-      "Commanders can now manage group members using the new Select feature in your groups",
-      "Easily book in and book out members, give and reward permissions to members and also mark members as 'ON-COURSE'",
-      "Members marked ON-COURSE will now no longer be automatically added into activities",
-      "Currently supports: Book In/Out, Make/Remove Admin, Remove from Group, Mark/Unmark as ON-COURSE",
-    ],
-    adminOnly: true,
-    date: "07/05/2024",
-    link: "https://social40.notion.site/v1-3-0-Select-Actions-70acf7745c114483abcc64d5c3703500",
-  },
-  "1.2.1": {
-    version: "1.2.1",
-    title: "Personal Statistics",
-    desc: "Added personal statistics tracking to display on profile",
-    updates: [
-      "Members can now add their own personal statistics to their profile",
-      "Commanders can also add for their soldiers to boost their motivation and monitor their progress",
-      "Currently supports IPPT, ATP, VOC and SOC",
-      "Members can earn Social40 points through these personal statistics",
-    ],
-    date: "06/05/2024",
-    link: "https://social40.notion.site/v1-2-1-Personal-Statistics-c520a2d031a34faab09d8596b2e15d0f?pvs=74",
-  },
-  "1.2.0": {
-    version: "1.2.0",
-    title: "COS - Points and Planning",
-    desc: "",
-    updates: [
-      "Created Social40 COS system which includes automatic points tracking",
-      "Plan your COS accurately with easy-to-use UI",
-      "Easily manage your monthly plan without worrying about keeping track of points",
-    ],
-    date: "05/05/2024",
-  },
-  "1.1.0": {
-    version: "1.1.0",
-    title: "Heat Acclimatisation (HA) Reports",
-    desc: "",
-    updates: [
-      "Easily calculate your soldiers HA currency",
-      "Say goodbye to manual tracking of HA",
-      "HA Reports can be downloaded into spreadsheets for other administrative uses if needed",
-    ],
-    date: "26/04/2024",
-  },
-  "1.0.0": {
-    version: "1.0.0",
-    title: "We are now out of Beta!",
-    desc: "You are part of Social40's beginning.",
-    updates: [
-      "After many updates and feedback, we have optimized this app to support 40SAR's relevant needs. We are thus happy to announce its first official release!",
-    ],
-    date: "21/04/2024",
-  },
-  "0.5.1": {
-    version: "0.5.1",
-    title: "Booking In",
-    desc: "You are part of Social40's beginning. Feel free to provide your feedback.",
-    updates: [
-      "Commanders can now mass book in soldiers via respective group pages",
-    ],
-    date: "04/03/2024",
-  },
-  "0.5.0": {
-    version: "0.5.0",
-    title: "Announcements",
-    desc: "You are part of Social40's beginning. Feel free to provide your feedback.",
-    updates: [
-      "Admins can now create posts to broadcast announcements",
-      "Friends feature has been removed",
-    ],
-    date: "27/02/2024",
-  },
-  "0.4.0": {
-    version: "0.4.0",
-    title: "Statuses Viewing",
-    desc: "You are part of Social40's beginning. Feel free to provide your feedback.",
-    updates: [
-      "Only active statuses will be shown on the main group page",
-      "A link is available to view all past statuses of every member in the group",
-    ],
-    date: "19/02/2024",
-  },
-  "0.3.3": {
-    version: "0.3.3",
-    title: "Profile Pictures",
-    desc: "You are part of Social40's beginning. Feel free to provide your feedback.",
-    updates: [
-      "Introducting Profile Pictures! Users can now add profile pictures to their Social40 accounts",
-    ],
-    date: "06/02/2024",
-  },
-  "0.3.2": {
-    version: "0.3.2",
-    title: "Download Activity Data",
-    desc: "You are part of Social40's beginning. Feel free to provide your feedback.",
-    updates: [
-      "Commanders can now download participation data from group activities into spreadsheets",
-    ],
-    date: "06/02/2024",
-  },
-  "0.3.1": {
-    version: "0.3.1",
-    title: "Strength and Participation",
-    desc: "You are part of Social40's beginning. Feel free to provide your feedback.",
-    updates: [
-      "Members will now be automatically added into activities upon creation except those on MC and statuses on the start date of the activity",
-      "Commanders can also opt out of the previous option and choose to custom-select who to join the activity",
-      "Commanders can now kick members from activities and categorise according to falling out",
-      "Commanders can more easily manage participation as present fall outs are already automatically not added",
-      "Commanders can view Strengths, MCs and Statuses distinctively",
-    ],
-    date: "24/01/2024",
-  },
-  "0.3.0": {
-    version: "0.3.0",
-    title: "Beta Release",
-    desc: "First stable beta release. Some features may still be unstable.",
-    updates: [],
-    date: "27/11/2023",
-  },
-  "0.2.3": {
-    version: "0.2.3",
-    title: "Activity Remarks",
-    desc: "Please note that this is still a beta testing version. Some features may still be unstable.",
-    updates: [
-      "Members can now make remarks to activities that have ended",
-      "Members must have been a participant of that activity",
-      "Activity owners can view all remarks made by participants",
-      "Serves to provide feedback to commanders regarding training",
-    ],
-    date: "19/11/2023",
-  },
-  "0.2.2": {
-    version: "0.2.2",
-    title: "Activity Feed for Groups",
-    desc: "Please note that this is still a beta testing version. Some features may still be unstable.",
-    updates: [
-      "Activities feed now can be split between groups and friends",
-      "Group activities created will be displayed on activity feed of group members",
-      "Group activities created can have restricitions (restrict to group members only and a deadline to join activity)",
-      "Members can request to participate in the activity, which will be tracked on their public profile",
-      "Their latest activities will be monitored to track their HA statuses",
-    ],
-    date: "16/11/2024",
-  },
-  "0.2.1": {
-    version: "0.2.1",
-    title: "Group Activities",
-    desc: "Please note that this is still a beta testing version. Some features may still be unstable.",
-    updates: [
-      "Group activities can now be added via the group dashboard by the group owner",
-    ],
-    date: "16/11/2024",
-  },
-  "0.1.2": {
-    version: "0.1.2",
-    title: "Specific BIBO Statuses",
-    desc: "Please note that this is still a beta testing version. Some features may still be unstable.",
-    updates: [
-      "COS, commanders and above can now view the BIBO statuses of members with a simple member ID search.",
-      "BIBO statuses can now be generated into excel sheets for easier references.",
-    ],
-    date: "07/11/2024",
-  },
-  "0.1.1": {
-    version: "0.1.1",
-    title: "Statuses",
-    desc: "Please note that this is still a beta testing version. Some features may still be unstable.",
-    updates: [
-      "Status feature are now available for both members and groups.",
-      "Commanders and above can now endorse statuses of others, NOT themselves.",
-    ],
-    date: "02/11/2024",
-  },
-  "0.1.0": {
-    version: "0.1.0",
-    title: "First Release",
-    desc: "Please note that this is still a beta testing version. Some features may still be unstable.",
-    updates: [],
-    date: "28/10/2024",
-  },
-} as {
-  [versionNumber: string]: {
-    version: string;
-    title: string;
-    desc: string;
-    updates: string[];
-    adminOnly?: boolean;
-    date: string;
-    link?: string;
-  };
-};
-
 export const GROUP_MEMBERS_SELECT_OPTIONS = {
   "Book In": {
     withOwner: true,
@@ -357,12 +157,6 @@ export const GROUP_ACTIVITY_CREATION_PROGRESS = [
   "Verifiying and adding participants...",
   "Finishing up...",
 ];
-
-export const VERSION_NUMBER = Object.keys(VERSION_MAP)[0] as string;
-export const VERSION_OBJ = VERSION_MAP[VERSION_NUMBER];
-export const VERSION_TITLE = VERSION_OBJ.title;
-export const VERSION_DESC = VERSION_OBJ.desc;
-export const VERSION_UPDATES = VERSION_OBJ.updates;
 
 export const MONTHS = [
   "JANUARY",
